# app/api/v1/admin/auth/verify-2fa/route.ts
// POST /api/v1/admin/auth/verify-2fa — step 2. 5 failed attempts invalidates
// the challengeToken (BR-3.6). Session ID re-minted on success (BR-3.8).
// See openapi-contract.yaml, docs/DESIGN-SYSTEM.md §5.

import { NextResponse } from "next/server";
import { createHash } from "node:crypto";
import bcrypt from "bcryptjs";
import { db } from "@/lib/db";
import { Verify2FAInputSchema } from "@/lib/schemas";
import { logActivity } from "@/lib/auth/activity-log";
import { verifyTotpOnce } from "@/lib/auth/totp";
import { createSessionCookieValue, SESSION_COOKIE_NAME, SESSION_COOKIE_OPTIONS } from "@/lib/auth/session";
import { crossSiteRefusal } from "@/lib/security/csrf";

const MAX_2FA_ATTEMPTS = 5; // BR-3.6

function errorResponse(code: string, message: string, status: number) {
  return NextResponse.json({ error: { code, message, details: null } }, { status });
}

function hashToken(rawToken: string): string {
  return createHash("sha256").update(rawToken).digest("hex");
}

function isTotpShaped(code: string): boolean {
  return /^\d{6}$/.test(code);
}

async function verifyAndConsumeRecoveryCode(adminId: string, hashedCodes: string[], code: string): Promise<boolean> {
  for (const hashed of hashedCodes) {
    if (await bcrypt.compare(code, hashed)) {
      await db.adminUser.update({
        where: { id: adminId },
        data: { recoveryCodes: hashedCodes.filter((c) => c !== hashed) },
      });
      return true;
    }
  }
  return false;
}

export async function POST(request: Request) {
  // BR-3.9 — login can't be driven from another site either.
  const refusal = crossSiteRefusal(request);
  if (refusal) return refusal;

  const body = await request.json().catch(() => null);
  const parsed = Verify2FAInputSchema.safeParse(body);
  if (!parsed.success) {
    return errorResponse("VALIDATION_ERROR", "Invalid request", 400);
  }

  const tokenHash = hashToken(parsed.data.challengeToken);
  const challenge = await db.loginChallenge.findUnique({
    where: { tokenHash },
    include: { adminUser: true },
  });

  // BR-3.5 — expired, already-consumed, or unknown all return the same
  // generic response as a wrong code; none of them restart the flow.
  if (!challenge || challenge.consumedAt || challenge.expiresAt.getTime() < Date.now()) {
    return errorResponse("CHALLENGE_EXPIRED", "Code is incorrect.", 401);
  }

  const isValid = isTotpShaped(parsed.data.code)
    ? challenge.adminUser.twoFactorSecret
      ? await verifyTotpOnce(db, challenge.adminUserId, challenge.adminUser.twoFactorSecret, parsed.data.code) // BR-3.14
      : false
    : await verifyAndConsumeRecoveryCode(challenge.adminUserId, challenge.adminUser.recoveryCodes, parsed.data.code);

  if (!isValid) {
    const attempts = challenge.attempts + 1;
    const exhausted = attempts >= MAX_2FA_ATTEMPTS;

    await db.loginChallenge.update({
      where: { id: challenge.id },
      data: {
        attempts,
        // BR-3.6 — invalidate the token itself once attempts are
        // exhausted, not just refuse further tries against it.
        consumedAt: exhausted ? new Date() : undefined,
      },
    });
    await logActivity({ adminUserId: challenge.adminUserId, action: "auth.2fa_failed", request });

    return errorResponse(
      "CHALLENGE_INVALID",
      isTotpShaped(parsed.data.code) ? "Code is incorrect." : "Recovery code is invalid.",
      401
    );
  }

  // Success — consume the challenge (BR-3.5, single-use regardless of
  // outcome) and mint a brand-new session, never reusing any pre-2FA
  // identifier (BR-3.8).
  await db.loginChallenge.update({ where: { id: challenge.id }, data: { consumedAt: new Date() } });
  const admin = await db.adminUser.update({
    where: { id: challenge.adminUserId },
    data: { lastLoginAt: new Date() },
    select: { sessionVersion: true },
  });
  await logActivity({ adminUserId: challenge.adminUserId, action: "auth.login", request });

  const sessionCookieValue = createSessionCookieValue(challenge.adminUserId, admin.sessionVersion);
  const response = NextResponse.json({
    sessionExpiresAt: new Date(Date.now() + SESSION_COOKIE_OPTIONS.maxAge * 1000).toISOString(),
  });
  response.cookies.set(SESSION_COOKIE_NAME, sessionCookieValue, SESSION_COOKIE_OPTIONS);
  return response;
}