# openapi-contract.yaml
openapi: 3.0.3
info:
  title: MALULEKE-KS Platform API
  version: "1.0.0"
  description: >
    Contract-first specification for the MALULEKE-KS platform.
    Derives from PLATFORM-CONSTITUTION-v1.md and BUSINESS-RULES-v1.md.
    Zod schemas are generated from this contract, never hand-written independently of it.
servers:
  - url: https://malulekeks.dev/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - {}
paths:
...
  /admin/auth/login:
    post:
      tags: [Auth]
      summary: Admin login step 1 — credentials
      description: On success, returns a challenge requiring 2FA (BR-3.1). Never issues a session on this call alone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Credentials accepted, 2FA challenge issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  challengeToken: { type: string }
        "401":
          description: Invalid credentials, or account locked (BR-3.2)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/auth/verify-2fa:
    post:
      tags: [Auth]
      summary: Admin login step 2 — TOTP or recovery code verification
      description: >
        `code` is either a 6-digit TOTP code or a recovery code (Design
        System §5). 5 failed attempts against the same challengeToken
        invalidates it (BR-3.6) rather than allowing continued guessing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeToken, code]
              properties:
                challengeToken: { type: string }
                code: { type: string, minLength: 6, maxLength: 32 }
      responses:
        "200":
          description: Session established
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionExpiresAt: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /admin/cv/experience/{id}:
    patch:
      tags: [CV (Admin)]
      summary: Update an experience entry
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExperienceInput" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/ExperienceEntry" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [CV (Admin)]
      summary: Delete an experience entry
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /admin/cv/education/{id}:
    patch:
      tags: [CV (Admin)]
      summary: Update an education entry
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EducationInput" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/EducationEntry" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [CV (Admin)]
      summary: Delete an education entry
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /admin/achievements/{id}:
    patch:
      tags: [Profile (Admin)]
      summary: Edit an achievement; publishing it puts it on the site and the CV
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AchievementInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Achievement" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Profile (Admin)]
      summary: Remove an achievement
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ============================================================
  # ADMIN — SETTINGS (Flags, Visitor Lenses)
  # ============================================================
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session
      description: >-
        HTTP-only, SameSite=Strict session cookie, issued only after 2FA verification (BR-3.1).
        Every state-changing admin request — and the two login steps — is refused with
        403 CSRF_REJECTED when its Origin or Sec-Fetch-Site names another site (BR-3.9).
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid admin session
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - NOT_FOUND
                - CLIENT_APPROVAL_REQUIRED
                - INVALID_STATUS_TRANSITION
                - RATE_LIMITED
                - ACCOUNT_LOCKED
                - CHALLENGE_EXPIRED
                - CHALLENGE_INVALID
                - LOOKUP_KEY_DEPRECATED
                - LOOKUP_KEY_EXISTS
                - OWNER_PERMISSION_REQUIRED
                - INTERNAL_ERROR
              description: Closed set, kept in lockstep with lib/schemas.ts ErrorCodeEnum — never a freeform string.
            message: { type: string }
            details: { type: object, nullable: true }
...
    ExperienceEntry:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        organization: { type: string }
        location: { type: string, nullable: true }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date, nullable: true }
        description: { type: string }
        highlights: { type: array, items: { type: string }, description: "CV achievement bullets, in order" }
        contentStatus: { type: string, enum: [draft, published, archived], description: "Only published roles appear on /cv and in the generated CV" }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — admin responses only. When published content goes live; null = as soon as published" }
        skills: { type: array, items: { type: string } }
    ExperienceInput:
      type: object
      required: [title, organization, startDate, description]
      properties:
        title: { type: string, minLength: 1 }
        organization: { type: string, minLength: 1 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date, nullable: true }
        location: { type: string, nullable: true }
        description: { type: string, minLength: 1 }
        highlights: { type: array, items: { type: string, maxLength: 300 }, maxItems: 15, default: [] }
        contentStatus: { type: string, enum: [draft, published, archived], description: "Omitted on create = published; omitted on update = unchanged" }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — schedule publication: with contentStatus published, the content goes live at this time (checked by every publish gate now). Omitted = unchanged; null = live as soon as published. Refused (400) on unpublished content" }
        skillIds: { type: array, items: { type: string }, default: [] }
    EducationEntry:
      type: object
      properties:
        id: { type: string }
        institution: { type: string }
        qualification: { type: string }
        fieldOfStudy: { type: string, nullable: true }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date, nullable: true, description: "null = still studying" }
        honors: { type: string, nullable: true }
        description: { type: string, nullable: true }
        certificateUrl: { type: string, nullable: true, description: "https only (database CHECK)" }
        contentStatus: { type: string, enum: [draft, published, archived], description: "Only published entries appear on /cv and in the generated CV" }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — admin responses only. When published content goes live; null = as soon as published" }
        skills: { type: array, items: { type: string }, description: "Skill names this qualification taught" }
        expectedGraduation: { type: string, format: date, nullable: true, description: "For a degree in progress" }
        coursework: { type: array, items: { type: string }, description: "Relevant modules" }
    EducationInput:
      type: object
      required: [institution, qualification, startDate]
      properties:
        institution: { type: string, minLength: 1 }
        qualification: { type: string, minLength: 1 }
        fieldOfStudy: { type: string, nullable: true }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date, nullable: true }
        honors: { type: string, nullable: true }
        description: { type: string, nullable: true }
        certificateUrl: { type: string, nullable: true, pattern: "^https://" }
        contentStatus: { type: string, enum: [draft, published, archived], description: "Omitted on create = published; omitted on update = unchanged" }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — schedule publication: with contentStatus published, the content goes live at this time (checked by every publish gate now). Omitted = unchanged; null = live as soon as published. Refused (400) on unpublished content" }
        skillIds: { type: array, items: { type: string }, default: [], description: "Replaces the entry's skill links" }
        expectedGraduation: { type: string, format: date, nullable: true, description: "Only for a degree in progress (no endDate) — database CHECK" }
        coursework: { type: array, items: { type: string, maxLength: 120 }, maxItems: 30, default: [] }
...
    Achievement:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        issuer: { type: string, nullable: true }
        achievedOn: { type: string, format: date }
        description: { type: string, nullable: true }
        url: { type: string, nullable: true }
        systemId: { type: string, nullable: true }
        contentStatus: { type: string, enum: [draft, published, archived] }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — admin responses only. When published content goes live; null = as soon as published" }
        sortOrder: { type: integer }
    AchievementInput:
      type: object
      required: [title, achievedOn]
      properties:
        title: { type: string, minLength: 1 }
        issuer: { type: string, nullable: true }
        achievedOn: { type: string, format: date }
        description: { type: string, nullable: true }
        url: { type: string, nullable: true, pattern: "^https://" }
        systemId: { type: string, nullable: true }
        contentStatus: { type: string, enum: [draft, published, archived], description: "Omitted on create = draft" }
        publishAt: { type: string, format: date-time, nullable: true, description: "BR-1.13 — schedule publication: with contentStatus published, the content goes live at this time (checked by every publish gate now). Omitted = unchanged; null = live as soon as published. Refused (400) on unpublished content" }
        sortOrder: { type: integer, minimum: 0, default: 0 }
...